requirements.txt
torch
torchvision
timm
albumentations
opencv-python
pandas
numpy
scikit-learn
matplotlib
seaborn
tqdm
pyyaml
